requirements.txt
# Inventory Forecasting POC - Requirements
# ============================================

# API/Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Time Series Forecasting
# Install Prophet (recommended for production)
# prophet>=1.1.0
# OR
# fbprophet>=0.7.0

# Fallback forecasting dependencies
statsmodels>=0.13.0

# Visualization (client-side chart rendering)
plotly>=5.15.0

# Optional: For better performance
# ujson>=5.7.0
# orjson>=3.8.0

